\name{requireGitHub_code}
\alias{requireGitHub_code}
\title{Print a requireGitHub declaration.}
\usage{
requireGitHub_code(...)
}
\arguments{
  \item{...}{unquoted package names.}
}
\value{
An invisible character vector of repository/package version
codes.
}
\description{
Print a requireGitHub declaration.
}
\examples{
requireGitHub_code(requireGitHub)
}
\author{
Toby Dylan Hocking
}

