% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{requireGitHub_package}
\alias{requireGitHub_package}
\title{require a package from GitHub}
\usage{
requireGitHub_package(username, path, sha1, pkg.name)
}
\arguments{
\item{username}{github username, e.g. "aldro61"}

\item{path}{repo name and path to R pkg, e.g. "mmit/Rpackage"}

\item{sha1}{hash}

\item{pkg.name}{the package name that you give to library ("mmit"
in the example above).}
}
\value{
was the package installed? TRUE or FALSE
}
\description{
require a specific version of a package from GitHub, installing
it if necessary.
}
\details{
This function can be used for packages which are not in the root
directory of a GitHub repo, or for packages which do not have the
same name as the GitHub repo.
}
\examples{
if(FALSE){
requireGitHub_package(
  "aldro61",
  "mmit/Rpackage",
  "edf81ba77fdd4b005ad89b81d9e12d289c8146e9",
  "mmit")
}
}
\author{
Toby Dylan Hocking
}

