\name{str_match_perl}
\alias{str_match_perl}
\title{Parse the first occurance of pattern from each of several strings
using (named) capturing regular expressions.}
\usage{
str_match_perl(string, pattern)
}
\arguments{
  \item{string}{character vector}

  \item{pattern}{character}
}
\value{
a matrix (with column names).
}
\description{
Parse the first occurance of pattern from each of several
strings using (named) capturing regular expressions.
}
\author{
Toby Dylan Hocking
}

