% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{requireGitHub}
\alias{requireGitHub}
\title{Install a specific package version from GitHub if it is not already.}
\usage{
requireGitHub(...)
}
\arguments{
\item{...}{character vectors matching username/repo@ref}
}
\value{
data.frame with 4 columns: GithubUsername,
GithubRepo, GithubSHA1, install (result of requireGitHub_package)
}
\description{
Install a specific package version from GitHub if it is not already.
}
\examples{
if(FALSE){
requireGitHub::requireGitHub(
  "tdhock/animint@057f34055ee876404caae7abc1c077a7bf126580",
  "rstudio/ggvis@8aa5ae207b3da0ff218fb5a3829bbdd59e54043c")
}
}
\author{
Toby Dylan Hocking
}

