% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{requireGitHub_code}
\alias{requireGitHub_code}
\title{Print a requireGitHub declaration.}
\usage{
requireGitHub_code(...)
}
\arguments{
\item{...}{unquoted package names.}
}
\value{
An invisible character vector of repository/package
version codes.
}
\description{
Print a requireGitHub declaration.
}
\examples{
if(FALSE){
requireGitHub_code(requireGitHub)
}
}
\author{
Toby Dylan Hocking
}
